Moved pdf viewer to delayed
