+++ scripts/delayed.js
@@ -1,4 +1,4 @@
-import { sampleRUM, fetchPlaceholders } from './lib-franklin.js';
+import { sampleRUM, fetchPlaceholders, getMetadata } from './lib-franklin.js';
 
 const placeholders = await fetchPlaceholders();
 // const isProd = window.location.hostname.endsWith(placeholders.hostname);
@@ -41,3 +41,35 @@ if (otId) {
   };
 }
 // OneTrust Cookies Consent Notice end
+
+// PDF Viewer for doc pages
+const docUrl = getMetadata('document-link');
+
+if (docUrl) {
+  const docFilename = docUrl.split('/').pop();
+
+  loadScript('https://documentservices.adobe.com/view-sdk/viewer.js');
+  let pdfAPIKey;
+  if (window.location.host.endsWith('.page') || window.location.host.startsWith('localhost')) {
+    pdfAPIKey = placeholders.pdfapikeypage;
+  } else if (window.location.host.endsWith('.live')) {
+    pdfAPIKey = placeholders.pdfapikeylive;
+  } else if (window.location.host.endsWith('merative.com')) {
+    pdfAPIKey = placeholders.pdfapikey;
+  }
+
+  if (pdfAPIKey) {
+    console.log(pdfAPIKey);
+    document.addEventListener('adobe_dc_view_sdk.ready', () => {
+      // eslint-disable-next-line no-undef
+      const adobeDCView = new AdobeDC.View({ clientId: pdfAPIKey, divId: 'adobe-dc-view' });
+      if (adobeDCView) {
+        // eslint-disable-next-line no-undef
+        adobeDCView.previewFile({
+          content: { location: { url: docUrl } },
+          metaData: { fileName: docFilename },
+        });
+      }
+    });
+  }
+}
